Check for error importing setuptools.command.test
In Setuptools 72.0.0, `setuptools.command.test` was temporarily removed.
This version of Setuptools was yanked, and 72.1.0 restored the
functionality (it appears that the removal didn't follow the standard
process, but this functionality has been deprecated for many years).

The functionality will be removed in a future version (probably very
soon) and that would break this `setup.py` as a result.

This commit, as a first step, handles the `ImportError` that results
when `setuptools.command.test` is removed, and only attempts to define
the `PyTest` command (the only thing depending on it) when it's
available.

For more details, please see:

https://github.com/pypa/setuptools/issues/4519

Although many projects were affected by the change, this one was the
unlucky "winner" of a popular Stack Overflow Q&A drawing attention to
the general issue:

https://stackoverflow.com/questions/78806100